arecasafe_backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
scikit-learn
joblib
tensorflow>=2.10
numpy
requests
python-dotenv
supabase
firebase-admin
Pillow
